+++ newsletter-bundle/CHANGELOG.md
@@ -2,4 +2,5 @@
 
 ## DEV
 
+ * Add a custom text field to the subscribe module (see contao/core-bundle#1512).
  * Use type="email" in the default template (see #20).